Organize textures a bit for better reuse
